Add reset test data to deactive user endpoint for tester role.

+++ auth-api/src/auth_api/services/reset.py
@@ -12,9 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Service for reset test data."""
+from typing import Dict
 
 from auth_api.models import User as UserModel
 from auth_api.models import db
+from auth_api.utils.roles import Role
 
 
 class ResetTestData:  # pylint:disable=too-few-public-methods
@@ -24,17 +26,18 @@ class ResetTestData:  # pylint:disable=too-few-public-methods
         """Return a reset test data service instance."""
 
     @staticmethod
-    def reset(token):
+    def reset(token_info: Dict):
         """Cleanup all the data from all tables create by the provided user id."""
-        user = UserModel.find_by_jwt_token(token)
-        if user:
-            # TODO need to find a way to avoid using protected function
-            for model_class in db.Model._decl_class_registry.values():  # pylint:disable=protected-access
-                if hasattr(model_class, 'created_by_id'):
-                    for model in model_class.query.filter_by(created_by_id=user.id).all():
-                        model.delete()
+        if Role.TESTER.value in token_info.get('realm_access').get('roles'):
+            user = UserModel.find_by_jwt_token(token_info)
+            if user:
+                # TODO need to find a way to avoid using protected function
+                for model_class in db.Model._decl_class_registry.values():  # pylint:disable=protected-access
+                    if hasattr(model_class, 'created_by_id'):
+                        for model in model_class.query.filter_by(created_by_id=user.id).all():
+                            model.delete()
 
-            user.modified_by = None
-            user.modified_by_id = None
-            db.session.delete(user)
-            db.session.commit()
+                user.modified_by = None
+                user.modified_by_id = None
+                db.session.delete(user)
+                db.session.commit()
